Issue #136: Moved reindexing out of the loop
